Add Exercise 1.38.
Actually, I like this! As I mentioned in my commit message, I would likely not
even have gone *looking for* an O(1) relation of this kind in a language like
Python. On that note, I've been sort of lucky that the sequences presented so
far have been "guessable"; I'm very curious to see if anybody else had a more
clever implementation. (I suspect something obvious, yet subtle, is lying in
wait here for the Dannys of the world.)