Add solutions up to 2.2.2.
Some interesting, if easy, exercises here. It's coming up a lot that Scheme has
a list representaiton which optimizes for *prepends* rather than *appends*,
which is honestly interesting. (Despite the fact that it might be obvious, since
a Scheme list is basically a singly-linked list with no tail reference.) For
example, it's not obvious how to do the `reverse` exercise in linear time.