Add Exercise 2.42.
Honestly, this took me a long time. I suspect it would have taken *much less*
time in Python. (It's not that hard a problem, really...) My solution doesn't
use any of this section's material, e.g. flatmap, which is disappointing. But my
solution depends on `and` and `or`, so I'm not sure how well these play with
accumulate, et al. since they're not proper *functions* -- recall they have a
special evaluation order -- and cannot be passed around. Seriously, try it!  You
can't do it!